修改了黑名单中的字段

+++ src/main/java/com/xlx/service/BlacklistServiceImpl.java
@@ -29,11 +29,22 @@ public class BlacklistServiceImpl implements BlacklistService {
     @Override
     public Boolean deleteOne(Blacklist blacklist) {
         try {
+            //如果记录存在，那么查找出来，然后减少累计
+            Blacklist queryByPhone = blacklistMapper.queryByPhone(blacklist);
+            if (queryByPhone==null)return true;
+            Long weights = queryByPhone.getWeights();
+            queryByPhone.setWeights(weights-1);
+            Object cc=weights-1==0?blacklistMapper.delete(queryByPhone):blacklistMapper.updateByPrimaryKey(queryByPhone);
             blacklistMapper.delete(blacklist);
             return true;
         } catch (Exception e) {
             e.printStackTrace();
         }
         return false;
+    }
+
+    @Override
+    public Blacklist queryByPhone(Blacklist blacklist) {
+        return blacklistMapper.queryByPhone(blacklist);
     }
 }
